requirements.txt
Flask
dnspython
python-whois
requests
pyOpenSSL
idna
gunicorn
